화면 모드 전환까지
