+++ app/src/main/AndroidManifest.xml
@@ -7,22 +7,31 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 
     <application
-        android:requestLegacyExternalStorage="true"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
+        android:requestLegacyExternalStorage="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.BootaCAM">
-        <activity android:name=".activity_camScreen"></activity>
-        <activity android:name=".activity_camList">
+        <activity android:name=".keyhash">
+        </activity>
+
+        <activity android:name=".ActivityCamScreen"
+            android:configChanges="keyboardHidden|orientation|screenSize">
+        </activity>
+
+        <activity android:name=".ActivityCamList"
+            android:configChanges="keyboardHidden|orientation|screenSize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity_camScreen"></activity>
+
+        <activity android:name=".MainActivity">
+        </activity>
     </application>
 
 </manifest>
