+++ app/src/main/res/layout/activity_camscreen.xml
@@ -1,17 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activity_camScreen"
+    tools:context=".ActivityCamScreen"
     android:padding="0dp">
 
     <com.google.android.exoplayer2.ui.PlayerView
-        android:id="@+id/video_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:id="@+id/playerview_exoplayer"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
         android:focusable="true" />
 
-</LinearLayout>
+    <ImageView
+        android:id="@+id/imageview_fullscreen"
+        android:layout_width="50dp"
+        android:layout_height="50dp"
+        android:src="@drawable/camscreen_fullscreen_40dp"
+        android:layout_gravity="right"/>
+
+</FrameLayout>
