+++ app/src/main/res/layout/item_camlist.xml
@@ -13,17 +13,17 @@
 
 
         <ImageView
-            android:id="@+id/recState"
+            android:id="@+id/imageview_recstate"
             android:layout_width="40dp"
             android:layout_height="40dp"
-            android:layout_weight="1"
+            android:layout_gravity="left"
             android:padding="3dp"
             android:src="@drawable/itemcamlist_recstop_40dp" />
 
 
         <TextView
-            android:id="@+id/camName"
-            android:layout_width="300dp"
+            android:id="@+id/textview_camname"
+            android:layout_width="wrap_content"
             android:layout_height="match_parent"
             android:layout_marginLeft="6dp"
             android:layout_weight="1"
@@ -34,10 +34,10 @@
             android:textSize="25dp" />
 
         <ImageView
-            android:id="@+id/camSetting"
+            android:id="@+id/imageview_setting"
             android:layout_width="40dp"
             android:layout_height="40dp"
-            android:layout_weight="1"
+
             android:src="@drawable/itemcamlist_setting_40dp" />
